Helper class added to wrap JwtBearerEvents
